requirements.txt
streamlit
openai>=1.30.0
python-dotenv
